+++ packages/toast/src/main.vue
@@ -34,11 +34,11 @@ export default {
   },
   data () {
     return {
-      forbidClick: false
+      forbidClick: false,
+      iconName: ''
     }
   },
   props: {
-    iconName: String,
     iconClass: String,
     msg: String,
     position: {
@@ -57,7 +57,8 @@ export default {
     iconSize: {
       type: String,
       default: '42px'
-    }
+    },
+    onClose: Function
   },
   computed: {
     customClass () {